requirements.txt
pandas
msal
requests
python-dotenv
openpyxl
openai
scikit-learn
fuzzywuzzy
python-Levenshtein